docs: all the lines
